feat!: Switch from Kong proxy to NGINIX proxy
BREAKING CHANGE: EdgeX 3.0 replaces the Kong reverse proxy
with a generic NGINX reverse proxy.  While the usage model
remains the same, there is a minor change to the method by
which a JWT authentication token is obtained.

This commit removes Kong and Postgres and replaces with a
much lighter weight NGINX and a new security-proxy-auth service.

Signed-off-by: Bryon Nevis <bryon.nevis@intel.com>